Phase 6 Step 5: zero-warning cleanup and strict build gates

+++ apps/daw_gui/audio/engine.cpp
@@ -19,7 +19,7 @@ void EngineShutdown(UiState& state) {
 
 bool EngineFillRealtimeBufferLocked(UiState& state, std::int16_t* outInterleaved, int frames, bool* reachedEnd) {
     if (state.project.projectSampleRate <= 0) {
-        std::fill(outInterleaved, outInterleaved + (frames * 2), 0);
+        std::fill(outInterleaved, outInterleaved + (frames * 2), static_cast<std::int16_t>(0));
         *reachedEnd = true;
         return false;
     }
@@ -34,7 +34,7 @@ bool EngineFillRealtimeBufferLocked(UiState& state, std::int16_t* outInterleaved
     const bool allowNoClipPlayback = runMetPlay || runMetRec || runCountInClick || (state.recording && state.inputMonitoring);
 
     if (state.project.clips.empty() && !allowNoClipPlayback) {
-        std::fill(outInterleaved, outInterleaved + (frames * 2), 0);
+        std::fill(outInterleaved, outInterleaved + (frames * 2), static_cast<std::int16_t>(0));
         *reachedEnd = true;
         return false;
     }
